Fix download_helper: avoid to download files when downloading more than one file Fix detail_provider: correctly encode entry for favs

+++ demo-app/lib/database/download_helper.dart
@@ -31,7 +31,9 @@ class DownloadsDB {
     final db = ObjectDB(FileSystemStorage(await getPath()));
     List val = await db.find({});
     for (var element in val) {
-      db.remove(element);
+      if (item["id"] == element["id"]) {
+        db.remove(element);
+      }
     }
     await db.close();
   }
